backend/requirements.txt
gunicorn
uvicorn

python-decouple

twilio
APScheduler

anyio==3.3.4
cffi==1.15.0
cryptography==35.0.0
dnspython==2.1.0
email-validator==1.1.3
fastapi==0.70.0
idna==3.3
Pillow==8.4.0
pybase64==1.2.0
pycparser==2.21
pydantic==1.8.2
PyJWT==2.3.0
pymongo==3.12.1
sniffio==1.2.0
starlette==0.16.0
typing_extensions==4.0.0
